Constructor refactoring
A container for the table and delete buttons has been created
to more correctly determine the selected row / cell / remove button.
Removed addEventListener from body and assigned to a separate container.
Instead of mouseover, I used mouseenter and mouseleave
to move the buttons around the table.
Redesigned the method of determining the coordinates
of the delete buttons.
As a result of refactoring, the visibleButtons method is removed
as unnecessary.
Removed unnecessary styling and the transmitted name for the container.